requirements_dev.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
codecov==2.0.15
coverage==4.4.1
Django==3.0.2
django-model-utils==4.0.0
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
idna==2.8
importlib-metadata==1.4.0
mccabe==0.6.1
mock==3.0.5
more-itertools==8.1.0
packaging==20.0
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
tox==3.14.3
urllib3==1.25.7
virtualenv==16.7.9
zipp==0.6.0
